Caching is now optional
